+++ backend/app-cvi-admin/build.gradle
@@ -2,5 +2,4 @@ bootJar { enabled = false }
 jar { enabled = true }
 
 dependencies {
-
 }
