+++ backend/domain-cvi-oauth-service/build.gradle
@@ -18,12 +18,12 @@ jacocoTestCoverageVerification { // 코드 커버리지 측정항목 시행
     violationRules { // 위반 규칙
         rule {
             enabled = true
-            element = 'CLASS'
+            element = 'BUNDLE'
 
             limit {
                 counter = 'INSTRUCTION'
                 value = 'COVEREDRATIO'
-                minimum = 0.80
+                minimum = 0.85
             }
 
             excludes = [
