+++ backend/domain-cvi-publicdata-service/src/test/java/com/cvi/CentralVaccinationInformationApplication.java
@@ -1,0 +1,12 @@
+package com.cvi;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CentralVaccinationInformationApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(CentralVaccinationInformationApplication.class, args);
+    }
+}
