+++ backend/domain-cvi-publicdata-service/src/test/java/com/cvi/service/PublicDataServiceTest.java
@@ -14,6 +14,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import org.mockito.BDDMockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
@@ -26,6 +27,7 @@ import java.util.stream.Stream;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.BDDMockito.*;
 import static org.mockito.BDDMockito.willReturn;
 import static org.mockito.Mockito.mock;
 
@@ -40,9 +42,6 @@ class PublicDataServiceTest {
     private PublicDataProperties publicDataProperties;
     @Autowired
     private PublicDataService publicDataService;
-
-    @MockBean
-    private PublicDataScheduler publicDataScheduler;
 
     private VaccinationParser vaccinationParser;
