+++ backend/domain-cvi-publicdata-service/src/test/resources/application.yml
@@ -1,0 +1,27 @@
+spring:
+  profiles:
+    active: test
+
+  datasource:
+    driver-class-name: org.h2.Driver
+    url: jdbc:h2:mem:test;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE;
+    username: sa
+    password:
+
+  jpa:
+    hibernate:
+      ddl-auto: create-drop
+    properties:
+      hibernate:
+        format_sql: true
+    generate-ddl: true
+
+security:
+  public:
+    data:
+      vaccination: publicdata-secret
+
+logging:
+  level:
+    org.hibernate.SQL: debug
+    org.hibernate.type: trace
