+++ backend/domain-cvi/src/test/resources/application.yml
@@ -1,0 +1,40 @@
+spring:
+  profiles:
+    active: test
+
+  datasource:
+    driver-class-name: org.h2.Driver
+    url: jdbc:h2:mem:test;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE;
+    username: sa
+    password:
+
+  h2:
+    console:
+      enabled: true
+
+  jpa:
+    hibernate:
+      ddl-auto: create-drop
+    properties:
+      hibernate:
+        format_sql: true
+    generate-ddl: true
+
+security:
+  jwt:
+    token:
+      secret-key: cvi_secret_key
+      expire-length: 3600000
+
+  auth:
+    naver:
+      client-secret: client-secret
+
+  public:
+    data:
+      vaccination: publicdata-secret
+
+logging:
+  level:
+    org.hibernate.SQL: debug
+    org.hibernate.type: trace
